requirements.txt
absl-py==2.4.0
attrs==26.1.0
cffi==2.0.0
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
flatbuffers==25.12.19
fonttools==4.62.1
kiwisolver==1.5.0
matplotlib==3.10.8
mediapipe==0.10.9
MouseInfo==0.1.3
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
packaging==26.0
pillow==12.2.0
protobuf==3.20.3
PyAutoGUI==0.9.54
pycparser==3.0
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyparsing==3.3.2
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
six==1.17.0
sounddevice==0.5.5
